rename user_task -> input_extraction
