+++ app-server/src/traces/mod.rs
@@ -2,6 +2,7 @@ pub mod consumer;
 pub mod data_plane_consumer;
 pub mod grpc_service;
 pub mod input_dedup;
+pub mod input_extraction;
 pub mod metadata;
 pub mod opentelemetry_json;
 #[cfg(feature = "signals")]
@@ -15,7 +16,6 @@ pub mod sampling;
 pub mod span_attributes;
 pub mod spans;
 pub mod tool_dedup;
-pub mod user_task;
 pub mod utils;
 
 pub const OBSERVATIONS_QUEUE: &str = "observations_queue";
